refactor(catalog): separate converter layer from service layer for clean architecture
Address code review feedback by introducing dedicated converter package that
separates OpenAPI ↔ DB conversion logic from the service layer, eliminating
architectural violations where service layer depended on presentation layer.

Key Changes:
- Created catalog/internal/converter package for OpenAPI ↔ DB conversions
- Moved all MCP conversion functions from service to converter package
- Relocated PropertyAccessor from service to converter package
- Removed OpenAPI imports from service layer (catalog/internal/db/service)
- Enhanced IsDuplicateKeyError to use vendor-specific error codes:
  * PostgreSQL: error code 23505 (unique_violation)
  * MySQL: error code 1062 (ER_DUP_ENTRY)
  * Replaces fragile string matching with reliable error code detection
