+++ catalog/internal/converter/property_accessor.go
@@ -1,4 +1,4 @@
-package service
+package converter
 
 import (
 	"encoding/json"
