+++ catalog/internal/db/models/mcp_server.go
